update benchmark dataset
vLLM uses the dataset found in https://huggingface.co/datasets/anon8231489123/ShareGPT_Vicuna_unfiltered/resolve/main/ShareGPT_V3_unfiltered_cleaned_split.json. In their version of the benchmark, they extract the first turn of human and gpt, then tokenize, then use the length of the tokenized sequence to prune requests for which the prompt and/or gpt response is too long.

It seemed strange to have to run the tokenizer every time we run the benchmark, so we decided to pre-filter the dataset by tokenizing (using Llama tokenizer) the first two turns of each conversation and permanently removing the conversations deemed too long/short. This is how ShareGPT_V3_filtered.json was generated.