update benchmark script and dataset
Benchmark changes:
There are now 2 benchmark scripts: benchmark_server and benchmark_server_zeus. The only difference between them is that benchmark_server_zeus also measures the energy by creating an zeus measurement window across the call to the benchmark function. Based on a run of Lllama-2-70b sharded across 4 nodes, the reported energy was 1679669 J by zeus and 1633937 J by accumulating TGI responses. The closeness of these measurements validate that the TGI instrumentation is working as intended.
The benchmark script also removes random sampling. Instead, it now runs all the prompts in the input file - so the input file should be small enough to account for this. For a request rate of 0.5, a dataset of 1000 inputs takes around 38 minutes.

Dataset changes:
Initially, we filtered by checking that num_prompt/completion_tokens < 1024. This has changed to < 800. Using 1024 lead to issues in TGI because the added chat template would increase the actual number of prompt tokens.
Also, we further sampled a dataset of 1000 prompts to use for benchmarking. We initially tried 2000, but found that it took too long (roughly double the number reported above). The filter_dataset file has been modified to include a way to randomly sample N inputs from a dataset.